# repository/swagger/v1/swagger.yml
openapi: 3.0.0
info:
  title: PPLDID API
  version: v1
  description: >-
     PPLDID provides a self-sustained environment for managing digital identifiers (DIDs). The ppld:did method links the identifier cryptographically to the DID Document and through also cryptographically linked provenance information in a public log it ensures resolving to the latest valid version of the DID Document.
  contact:
    email: jerry.zhang
  license:
    name: MIT License
    url: https://opensource.org/licenses/MIT
servers:
- url: https://ppldid.peopledata.org.cn/
  description: Official production system hosted by Peopledata.
security:
  - BearerAuth: []
paths:
...
  /1.0/identifiers/{identifier}:
    get:
      summary: Resolve a DID.
      tags:
        - Uniresolver Interface
      description:
        <p>This endpoint resolves a DID. As input it takes the DID, plus DID resolution
        <a href="https://www.w3.org/TR/did-core/#did-resolution-options">options</a>. The output is a DID
        document in one of the supported representations, plus metadata.</p>
        <p>See the <a href="https://w3c-ccg.github.io/did-resolution/">DID Resolution</a>
        specification for additional details.</p>
      operationId: resolve
      parameters:
        - in: path
          required: true
          name: identifier
          schema:
            type: string
          description: The DID to be resolved.
          examples:
            example1:
              value: did:ppld:zQmaBZTghndXTgxNwfbdpVLWdFf6faYE4oeuN2zzXdQt1kh
              description: A DID using the `ppld` method.
        - in: header
          required: false
          name: Accept
          schema:
            type: string
          description: The requested media type of the DID document representation or DID resolution result. See <a href="https://www.w3.org/TR/did-core/#representations">https://www.w3.org/TR/did-core/#representations</a> and <a href="https://w3c-ccg.github.io/did-resolution/#did-resolution-result">https://w3c-ccg.github.io/did-resolution/#did-resolution-result</a>.
          examples:
            application/did+json:
              value: application/json
              description: Media type of a DID document (JSON representation).
            application/did+ld+json:
              value: application/did+ld+json
              description: Media type of a DID document (JSON-LD representation).
            application/ld+json;profile="https://w3id.org/did-resolution":
              value: application/ld+json;profile="https://w3id.org/did-resolution"
              description: Media type of a DID resolution result (JSON-LD representation).
      responses:
        '200':
          description: successfully resolved!
          content:
            application/did+json:
              schema:
                type: object
                description: The DID document (JSON representation).
            application/did+ld+json:
              schema:
                type: object
                description: The DID document (JSON-LD representation).
            application/ld+json;profile="https://w3id.org/did-resolution":
              schema:
                $ref: '#/components/schemas/ResolutionResult'
        '400':
          description: invalid input!
        '410':
          description: successfully resolved (deactivated)!
          content:
            application/did+json:
              schema:
                type: object
                description: The deactivated DID document (JSON representation).
            application/did+ld+json:
              schema:
                type: object
                description: The deactivated DID document (JSON-LD representation).
            application/ld+json;profile="https://w3id.org/did-resolution":
              schema:
                $ref: '#/components/schemas/ResolutionResult'
        '500':
          description: error!
...
components:
  schemas:
...
    ResolutionResult:
      description: The DID resolution result.
      type: object
      additionalProperties: false
      properties:
        didDocument:
          type: object
        didResolutionMetadata:
          type: object
        didDocumentMetadata:
          type: object
...